requirements.txt
# Python 3.13 Compatible Versions
pymongo==4.6.0
pandas==2.2.3
python-dotenv==1.0.0
matplotlib==3.8.0
seaborn==0.13.0
numpy==1.26.4
